Module aws_sdk_cloudsearch::error

Modules

base_exception

See BaseException

disabled_operation_exception

See DisabledOperationException

internal_exception

See InternalException

invalid_type_exception

See InvalidTypeException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

validation_exception

See ValidationException

Structs

BaseException

An error occurred while processing the request.

BuildSuggestersError

CreateDomainError

DefineAnalysisSchemeError

DefineExpressionError

DefineIndexFieldError

DefineSuggesterError

DeleteAnalysisSchemeError

DeleteDomainError

DeleteExpressionError

DeleteIndexFieldError

DeleteSuggesterError

DescribeAnalysisSchemesError

DescribeAvailabilityOptionsError

DescribeDomainEndpointOptionsError

DescribeDomainsError

DescribeExpressionsError

DescribeIndexFieldsError

DescribeScalingParametersError

DescribeServiceAccessPoliciesError

DescribeSuggestersError

DisabledOperationException

The request was rejected because it attempted an operation which is not enabled.

IndexDocumentsError

InternalException

An internal error occurred while processing the request. If this problem persists,

InvalidTypeException

The request was rejected because it specified an invalid type definition.

LimitExceededException

The request was rejected because a resource limit has already been met.

ListDomainNamesError

ResourceAlreadyExistsException

The request was rejected because it attempted to create a resource that already exists.

ResourceNotFoundException

The request was rejected because it attempted to reference a resource that does not exist.

UpdateAvailabilityOptionsError

UpdateDomainEndpointOptionsError

UpdateScalingParametersError

UpdateServiceAccessPoliciesError

ValidationException

The request was rejected because it has invalid parameters.

Enums

BuildSuggestersErrorKind

CreateDomainErrorKind

DefineAnalysisSchemeErrorKind

DefineExpressionErrorKind

DefineIndexFieldErrorKind

DefineSuggesterErrorKind

DeleteAnalysisSchemeErrorKind

DeleteDomainErrorKind

DeleteExpressionErrorKind

DeleteIndexFieldErrorKind

DeleteSuggesterErrorKind

DescribeAnalysisSchemesErrorKind

DescribeAvailabilityOptionsErrorKind

DescribeDomainEndpointOptionsErrorKind

DescribeDomainsErrorKind

DescribeExpressionsErrorKind

DescribeIndexFieldsErrorKind

DescribeScalingParametersErrorKind

DescribeServiceAccessPoliciesErrorKind

DescribeSuggestersErrorKind

IndexDocumentsErrorKind

ListDomainNamesErrorKind

UpdateAvailabilityOptionsErrorKind

UpdateDomainEndpointOptionsErrorKind

UpdateScalingParametersErrorKind

UpdateServiceAccessPoliciesErrorKind