Module aws_sdk_cloudsearch::input

Modules

build_suggesters_input

See BuildSuggestersInput

create_domain_input

See CreateDomainInput

define_analysis_scheme_input

See DefineAnalysisSchemeInput

define_expression_input

See DefineExpressionInput

define_index_field_input

See DefineIndexFieldInput

define_suggester_input

See DefineSuggesterInput

delete_analysis_scheme_input

See DeleteAnalysisSchemeInput

delete_domain_input

See DeleteDomainInput

delete_expression_input

See DeleteExpressionInput

delete_index_field_input

See DeleteIndexFieldInput

delete_suggester_input

See DeleteSuggesterInput

describe_analysis_schemes_input

See DescribeAnalysisSchemesInput

describe_availability_options_input

See DescribeAvailabilityOptionsInput

describe_domain_endpoint_options_input

See DescribeDomainEndpointOptionsInput

describe_domains_input

See DescribeDomainsInput

describe_expressions_input

See DescribeExpressionsInput

describe_index_fields_input

See DescribeIndexFieldsInput

describe_scaling_parameters_input

See DescribeScalingParametersInput

describe_service_access_policies_input

See DescribeServiceAccessPoliciesInput

describe_suggesters_input

See DescribeSuggestersInput

index_documents_input

See IndexDocumentsInput

list_domain_names_input

See ListDomainNamesInput

update_availability_options_input

See UpdateAvailabilityOptionsInput

update_domain_endpoint_options_input

See UpdateDomainEndpointOptionsInput

update_scaling_parameters_input

See UpdateScalingParametersInput

update_service_access_policies_input

See UpdateServiceAccessPoliciesInput

Structs

BuildSuggestersInput

Container for the parameters to the BuildSuggester operation. Specifies the name of the domain you want to update.

CreateDomainInput

Container for the parameters to the CreateDomain operation. Specifies a name for the new search domain.

DefineAnalysisSchemeInput

Container for the parameters to the DefineAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme configuration.

DefineExpressionInput

Container for the parameters to the DefineExpression operation. Specifies the name of the domain you want to update and the expression you want to configure.

DefineIndexFieldInput

Container for the parameters to the DefineIndexField operation. Specifies the name of the domain you want to update and the index field configuration.

DefineSuggesterInput

Container for the parameters to the DefineSuggester operation. Specifies the name of the domain you want to update and the suggester configuration.

DeleteAnalysisSchemeInput

Container for the parameters to the DeleteAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme you want to delete.

DeleteDomainInput

Container for the parameters to the DeleteDomain operation. Specifies the name of the domain you want to delete.

DeleteExpressionInput

Container for the parameters to the DeleteExpression operation. Specifies the name of the domain you want to update and the name of the expression you want to delete.

DeleteIndexFieldInput

Container for the parameters to the DeleteIndexField operation. Specifies the name of the domain you want to update and the name of the index field you want to delete.

DeleteSuggesterInput

Container for the parameters to the DeleteSuggester operation. Specifies the name of the domain you want to update and name of the suggester you want to delete.

DescribeAnalysisSchemesInput

Container for the parameters to the DescribeAnalysisSchemes operation. Specifies the name of the domain you want to describe. To limit the response to particular analysis schemes, specify the names of the analysis schemes you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeAvailabilityOptionsInput

Container for the parameters to the DescribeAvailabilityOptions operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeDomainEndpointOptionsInput

Container for the parameters to the DescribeDomainEndpointOptions operation. Specify the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeDomainsInput

Container for the parameters to the DescribeDomains operation. By default shows the status of all domains. To restrict the response to particular domains, specify the names of the domains you want to describe.

DescribeExpressionsInput

Container for the parameters to the DescribeDomains operation.

DescribeIndexFieldsInput

Container for the parameters to the DescribeIndexFields operation. Specifies the name of the domain you want to describe. To restrict the response to particular index fields, specify the names of the index fields you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeScalingParametersInput

Container for the parameters to the DescribeScalingParameters operation. Specifies the name of the domain you want to describe.

DescribeServiceAccessPoliciesInput

Container for the parameters to the DescribeServiceAccessPolicies operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

DescribeSuggestersInput

Container for the parameters to the DescribeSuggester operation. Specifies the name of the domain you want to describe. To restrict the response to particular suggesters, specify the names of the suggesters you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

IndexDocumentsInput

Container for the parameters to the IndexDocuments operation. Specifies the name of the domain you want to re-index.

ListDomainNamesInput

UpdateAvailabilityOptionsInput

Container for the parameters to the UpdateAvailabilityOptions operation. Specifies the name of the domain you want to update and the Multi-AZ availability option.

UpdateDomainEndpointOptionsInput

Container for the parameters to the UpdateDomainEndpointOptions operation. Specifies the name of the domain you want to update and the domain endpoint options.

UpdateScalingParametersInput

Container for the parameters to the UpdateScalingParameters operation. Specifies the name of the domain you want to update and the scaling parameters you want to configure.

UpdateServiceAccessPoliciesInput

Container for the parameters to the UpdateServiceAccessPolicies operation. Specifies the name of the domain you want to update and the access rules you want to configure.