Module aws_sdk_cloudsearch::model

Modules

access_policies_status

See AccessPoliciesStatus

analysis_options

See AnalysisOptions

analysis_scheme

See AnalysisScheme

analysis_scheme_status

See AnalysisSchemeStatus

availability_options_status

See AvailabilityOptionsStatus

date_array_options

See DateArrayOptions

date_options

See DateOptions

document_suggester_options

See DocumentSuggesterOptions

domain_endpoint_options

See DomainEndpointOptions

domain_endpoint_options_status

See DomainEndpointOptionsStatus

domain_status

See DomainStatus

double_array_options

See DoubleArrayOptions

double_options

See DoubleOptions

expression

See Expression

expression_status

See ExpressionStatus

index_field

See IndexField

index_field_status

See IndexFieldStatus

int_array_options

See IntArrayOptions

int_options

See IntOptions

lat_lon_options

See LatLonOptions

limits

See Limits

literal_array_options

See LiteralArrayOptions

literal_options

See LiteralOptions

option_status

See OptionStatus

scaling_parameters

See ScalingParameters

scaling_parameters_status

See ScalingParametersStatus

service_endpoint

See ServiceEndpoint

suggester

See Suggester

suggester_status

See SuggesterStatus

text_array_options

See TextArrayOptions

text_options

See TextOptions

Structs

AccessPoliciesStatus

The configured access rules for the domain's document and search endpoints, and the current status of those rules.

AnalysisOptions

Synonyms, stopwords, and stemming options for an analysis scheme. Includes tokenization dictionary for Japanese.

AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, JapaneseTokenizationDictionary and AlgorithmicStemming.

AnalysisSchemeStatus

The status and configuration of an AnalysisScheme.

AvailabilityOptionsStatus

The status and configuration of the domain's availability options.

DateArrayOptions

Options for a field that contains an array of dates. Present if IndexFieldType specifies the field is of type date-array. All options are enabled by default.

DateOptions

Options for a date field. Dates and times are specified in UTC (Coordinated Universal Time) according to IETF RFC3339: yyyy-mm-ddT00:00:00Z. Present if IndexFieldType specifies the field is of type date. All options are enabled by default.

DocumentSuggesterOptions

Options for a search suggester.

DomainEndpointOptions

The domain's endpoint options.

DomainEndpointOptionsStatus

The configuration and status of the domain's endpoint options.

DomainStatus

The current status of the search domain.

DoubleArrayOptions

Options for a field that contains an array of double-precision 64-bit floating point values. Present if IndexFieldType specifies the field is of type double-array. All options are enabled by default.

DoubleOptions

Options for a double-precision 64-bit floating point field. Present if IndexFieldType specifies the field is of type double. All options are enabled by default.

Expression

A named expression that can be evaluated at search time. Can be used to sort the search results, define other expressions, or return computed information in the search results.

ExpressionStatus

The value of an Expression and its current status.

IndexField

Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType.

IndexFieldStatus

The value of an IndexField and its current status.

IntArrayOptions

Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array. All options are enabled by default.

IntOptions

Options for a 64-bit signed integer field. Present if IndexFieldType specifies the field is of type int. All options are enabled by default.

LatLonOptions

Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon. All options are enabled by default.

Limits

LiteralArrayOptions

Options for a field that contains an array of literal strings. Present if IndexFieldType specifies the field is of type literal-array. All options are enabled by default.

LiteralOptions

Options for literal field. Present if IndexFieldType specifies the field is of type literal. All options are enabled by default.

OptionStatus

The status of domain configuration option.

ScalingParameters

The desired instance type and desired number of replicas of each index partition.

ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.

ServiceEndpoint

The endpoint to which service requests can be submitted.

Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

SuggesterStatus

The value of a Suggester and its current status.

TextArrayOptions

Options for a field that contains an array of text strings. Present if IndexFieldType specifies the field is of type text-array. A text-array field is always searchable. All options are enabled by default.

TextOptions

Options for text field. Present if IndexFieldType specifies the field is of type text. A text field is always searchable. All options are enabled by default.

Enums

AlgorithmicStemming

AnalysisSchemeLanguage

An IETF RFC 4646 language code or mul for multiple languages.

IndexFieldType

The type of field. The valid options for a field depend on the field type. For more information about the supported field types, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

OptionState

The state of processing a change to an option. One of:

PartitionInstanceType

The instance type (such as search.m1.small) on which an index partition is hosted.

SuggesterFuzzyMatching

TlsSecurityPolicy

The minimum required TLS version.