Module aws_sdk_cloudsearch::output

Modules

build_suggesters_output

See BuildSuggestersOutput

create_domain_output

See CreateDomainOutput

define_analysis_scheme_output

See DefineAnalysisSchemeOutput

define_expression_output

See DefineExpressionOutput

define_index_field_output

See DefineIndexFieldOutput

define_suggester_output

See DefineSuggesterOutput

delete_analysis_scheme_output

See DeleteAnalysisSchemeOutput

delete_domain_output

See DeleteDomainOutput

delete_expression_output

See DeleteExpressionOutput

delete_index_field_output

See DeleteIndexFieldOutput

delete_suggester_output

See DeleteSuggesterOutput

describe_analysis_schemes_output

See DescribeAnalysisSchemesOutput

describe_availability_options_output

See DescribeAvailabilityOptionsOutput

describe_domain_endpoint_options_output

See DescribeDomainEndpointOptionsOutput

describe_domains_output

See DescribeDomainsOutput

describe_expressions_output

See DescribeExpressionsOutput

describe_index_fields_output

See DescribeIndexFieldsOutput

describe_scaling_parameters_output

See DescribeScalingParametersOutput

describe_service_access_policies_output

See DescribeServiceAccessPoliciesOutput

describe_suggesters_output

See DescribeSuggestersOutput

index_documents_output

See IndexDocumentsOutput

list_domain_names_output

See ListDomainNamesOutput

update_availability_options_output

See UpdateAvailabilityOptionsOutput

update_domain_endpoint_options_output

See UpdateDomainEndpointOptionsOutput

update_scaling_parameters_output

See UpdateScalingParametersOutput

update_service_access_policies_output

See UpdateServiceAccessPoliciesOutput

Structs

BuildSuggestersOutput

The result of a BuildSuggester request. Contains a list of the fields used for suggestions.

CreateDomainOutput

The result of a CreateDomainRequest. Contains the status of a newly created domain.

DefineAnalysisSchemeOutput

The result of a DefineAnalysisScheme request. Contains the status of the newly-configured analysis scheme.

DefineExpressionOutput

The result of a DefineExpression request. Contains the status of the newly-configured expression.

DefineIndexFieldOutput

The result of a DefineIndexField request. Contains the status of the newly-configured index field.

DefineSuggesterOutput

The result of a DefineSuggester request. Contains the status of the newly-configured suggester.

DeleteAnalysisSchemeOutput

The result of a DeleteAnalysisScheme request. Contains the status of the deleted analysis scheme.

DeleteDomainOutput

The result of a DeleteDomain request. Contains the status of a newly deleted domain, or no status if the domain has already been completely deleted.

DeleteExpressionOutput

The result of a DeleteExpression request. Specifies the expression being deleted.

DeleteIndexFieldOutput

The result of a DeleteIndexField request.

DeleteSuggesterOutput

The result of a DeleteSuggester request. Contains the status of the deleted suggester.

DescribeAnalysisSchemesOutput

The result of a DescribeAnalysisSchemes request. Contains the analysis schemes configured for the domain specified in the request.

DescribeAvailabilityOptionsOutput

The result of a DescribeAvailabilityOptions request. Indicates whether or not the Multi-AZ option is enabled for the domain specified in the request.

DescribeDomainEndpointOptionsOutput

The result of a DescribeDomainEndpointOptions request. Contains the status and configuration of a search domain's endpoint options.

DescribeDomainsOutput

The result of a DescribeDomains request. Contains the status of the domains specified in the request or all domains owned by the account.

DescribeExpressionsOutput

The result of a DescribeExpressions request. Contains the expressions configured for the domain specified in the request.

DescribeIndexFieldsOutput

The result of a DescribeIndexFields request. Contains the index fields configured for the domain specified in the request.

DescribeScalingParametersOutput

The result of a DescribeScalingParameters request. Contains the scaling parameters configured for the domain specified in the request.

DescribeServiceAccessPoliciesOutput

The result of a DescribeServiceAccessPolicies request.

DescribeSuggestersOutput

The result of a DescribeSuggesters request.

IndexDocumentsOutput

The result of an IndexDocuments request. Contains the status of the indexing operation, including the fields being indexed.

ListDomainNamesOutput

The result of a ListDomainNames request. Contains a list of the domains owned by an account.

UpdateAvailabilityOptionsOutput

The result of a UpdateAvailabilityOptions request. Contains the status of the domain's availability options.

UpdateDomainEndpointOptionsOutput

The result of a UpdateDomainEndpointOptions request. Contains the configuration and status of the domain's endpoint options.

UpdateScalingParametersOutput

The result of a UpdateScalingParameters request. Contains the status of the newly-configured scaling parameters.

UpdateServiceAccessPoliciesOutput

The result of an UpdateServiceAccessPolicies request. Contains the new access policies.